refactor(ecs): move system to system dir
