+++ engine/include/cubos/engine/cubos.hpp
@@ -9,7 +9,7 @@
 
 #include <cubos/core/ecs/dispatcher.hpp>
 #include <cubos/core/ecs/event_pipe.hpp>
-#include <cubos/core/ecs/system.hpp>
+#include <cubos/core/ecs/system/system.hpp>
 #include <cubos/core/ecs/world.hpp>
 
 namespace cubos::engine
